# api-server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.createUserRequest:
    properties:
      email:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      password:
        type: string
      plan_id:
        type: string
    type: object
  handlers.createUserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
  handlers.deviceItem:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      xray_uuid:
        type: string
    type: object
...
  handlers.updateUserRequest:
    properties:
      is_active:
        type: boolean
      name:
        type: string
      plan_expires_at:
        type: string
      plan_id:
        type: string
      plan_started_at:
        type: string
      status:
        type: string
    type: object
...
  handlers.userDetailResponse:
    properties:
      created_at:
        type: string
      devices:
        items:
          $ref: '#/definitions/handlers.deviceItem'
        type: array
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      plan_expires_at:
        type: string
      plan_id:
        type: string
      plan_name:
        type: string
      plan_started_at:
        type: string
      status:
        type: string
      sub_token:
        type: string
      traffic_limit:
        type: integer
      traffic_reset_at:
        type: string
      traffic_used:
        type: integer
    type: object
  handlers.userListItem:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      plan_expires_at:
        type: string
      plan_id:
        type: string
      plan_name:
        type: string
      status:
        type: string
      traffic_limit:
        type: integer
      traffic_used:
        type: integer
    type: object
  handlers.userListResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/handlers.userListItem'
        type: array
    type: object
...
  handlers.userTrafficResponse:
    properties:
      by_node:
        items:
          $ref: '#/definitions/services.UserNodeTraffic'
        type: array
      percentage:
        type: number
      traffic_limit:
        type: integer
      traffic_remaining:
        type: integer
      traffic_used:
        type: integer
      unlimited:
        type: boolean
      window:
        type: string
    type: object
...
  services.LoginEntry:
    properties:
      actor_type:
        type: string
      attempted_email:
        type: string
      created_at:
        type: string
      failure_reason:
        type: string
      id:
        type: string
      ip:
        type: string
      success:
        type: boolean
      user_agent:
        type: string
      user_id:
        type: string
    type: object
...
  services.UserNodeTraffic:
    properties:
      download:
        type: integer
      node_id:
        type: string
      node_name:
        type: string
      total:
        type: integer
      upload:
        type: integer
    type: object
host: localhost:2053
info:
  contact: {}
  description: VPN Panel Management API
  title: Proxima VPN API
  version: 0.1.0
paths:
...
  /admin/users:
    get:
      description: Returns paginated users with optional search and status filter
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by email or name
        in: query
        name: search
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.userListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - admin-users
    post:
      consumes:
      - application/json
      description: Creates a new user account with active status
      parameters:
      - description: User details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.createUserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - admin-users
  /admin/users/{id}:
    delete:
      description: Soft-deactivates a user (sets status to suspended)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - admin-users
    get:
      description: Returns a single user with plan and device details
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.userDetailResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user
      tags:
      - admin-users
    put:
      consumes:
      - application/json
      description: Partially updates a user's status, plan, or active state
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.updateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - admin-users
  /admin/users/{id}/login-history:
    get:
      description: Returns successful and failed login attempts for one user, newest
        first
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Maximum attempts to return (default 50, max 200)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.LoginEntry'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user login history
      tags:
      - admin-users
  /admin/users/{id}/reset-traffic:
    post:
      description: Resets traffic_used to 0 and sets traffic_reset_at to NOW()
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reset user traffic
      tags:
      - admin-users
  /admin/users/{id}/traffic:
    get:
      description: Returns quota, usage, remaining traffic and a per-node breakdown
        for one user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: today, week, or month (default month)
        in: query
        name: window
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.userTrafficResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user traffic
      tags:
      - admin-users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"